разбиение файла main.js на модули
